requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
Flask_HTTPAuth==4.7.0
k_means_constrained==0.7.0
matplotlib==3.5.3
modAL==0.4.1
numpy==1.21.6
pandas==1.3.5
Pillow==9.2.0
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.13.0
selenium==4.14.0
statsmodels==0.14.0
Werkzeug==2.2.3
xgboost==1.6.2
itu-p1203==1.8.3
